Run the simulation at 30 Hz on the device
Measured on hardware at 60 Hz: 21.5 fps, with the frame spending 19.9 ms
of its 46.5 ms on three ticks, 10.4 on beginFrame and 9.2 rasterizing.
CPU was 39.4 ms against a 14.75 ms DMA floor, so the frame rate is set by
this core and half of that is the simulation.

TICK_RATE becomes a macro the platform can override. The constants in
sim/config.hpp are all derived from per-second values and were originally
tuned at 30 Hz, so this is the rate they were written for. Solving the
measured cost model for 30 Hz gives about 34 ms a frame, near 29 fps.

Checked that the game still plays the same over 9 seeds x 3 weapons x 120
seconds: the hit rate, the feeding rate and the death rate are unchanged,
and kills are 8% higher. The hit rate matters most -- it shows the laser,
which covers 8 FU per tick at 30 Hz, is not tunnelling through targets.

ctest now runs the suite at both rates. That immediately caught a test
that counted ticks rather than seconds, so at 30 Hz it simulated twice
the game time and the player was absorbed before the assertion. The
simulation itself matches to the millisecond across rates.

Also casts the score values for %u: uint32_t is unsigned long on 32-bit
ARM, which the host build never showed.

+++ core/src/render/hud.cpp
@@ -116,10 +116,10 @@ int upgradeIconPolygon(int kind, int cx, int cy, g2::vec2i *pts, int scale8) {
     }
     case sim::UpgradeKind::EXTRA_CORE: {  // four-pointed star
       const g2::vec2i p[8] = {
-          {cx, cy - sc(7, s)},           {cx + sc(2, s), cy - sc(2, s)},
-          {cx + sc(7, s), cy},           {cx + sc(2, s), cy + sc(2, s)},
-          {cx, cy + sc(7, s)},           {cx - sc(2, s), cy + sc(2, s)},
-          {cx - sc(7, s), cy},           {cx - sc(2, s), cy - sc(2, s)}};
+          {cx, cy - sc(7, s)}, {cx + sc(2, s), cy - sc(2, s)},
+          {cx + sc(7, s), cy}, {cx + sc(2, s), cy + sc(2, s)},
+          {cx, cy + sc(7, s)}, {cx - sc(2, s), cy + sc(2, s)},
+          {cx - sc(7, s), cy}, {cx - sc(2, s), cy - sc(2, s)}};
       for (int i = 0; i < 8; i++) pts[i] = p[i];
       return 8;
     }
@@ -135,8 +135,7 @@ void drawUpgradeIcon(g2::Graphics2D &g, int kind, int cx, int cy, g2::Color c,
 }
 
 static g2::Color dimmed(g2::Color c) {
-  return g2::makeColor(g2::colorR(c) / 3, g2::colorG(c) / 3,
-                       g2::colorB(c) / 3);
+  return g2::makeColor(g2::colorR(c) / 3, g2::colorG(c) / 3, g2::colorB(c) / 3);
 }
 
 // Upgrades (bottom left): icon + level pips; spare cores (bottom right).
@@ -203,8 +202,9 @@ void Renderer::drawHud(g2::Graphics2D &g, int oy) {
       }
       setHudFont(g, HudFont::SMALL);
       if (game.highScore() > 0) {
-        std::snprintf(buf, sizeof(buf), "HIGH SCORE %u", game.highScore());
-        std::snprintf(alt, sizeof(alt), "HI %u", game.highScore());
+        std::snprintf(buf, sizeof(buf), "HIGH SCORE %u",
+                      (unsigned)game.highScore());
+        std::snprintf(alt, sizeof(alt), "HI %u", (unsigned)game.highScore());
         drawCenteredFit(g, oy + uiY(236), buf, alt, HUD_DIM);
       }
       int lineH = g.lineAdvance() + ui(6, 2);
@@ -296,7 +296,7 @@ void Renderer::drawHud(g2::Graphics2D &g, int oy) {
 
       // Score (top center): the shown value rolls up towards the real one
       setHudFont(g, HudFont::MEDIUM);
-      std::snprintf(buf, sizeof(buf), "%u", scoreShown());
+      std::snprintf(buf, sizeof(buf), "%u", (unsigned)scoreShown());
       {
         int tw = g.measureText(buf);
         int sh = ui_.fontMult;
@@ -359,11 +359,12 @@ void Renderer::drawHud(g2::Graphics2D &g, int oy) {
         drawCenteredFit(g, oy + uiY(100), "YOU WERE DEVOURED", "DEVOURED",
                         g2::makeColor(255, 90, 90));
         setHudFont(g, ui_.compact ? HudFont::SMALL : HudFont::MEDIUM);
-        std::snprintf(buf, sizeof(buf), "SCORE %u", game.score());
-        std::snprintf(alt, sizeof(alt), "%u", game.score());
+        std::snprintf(buf, sizeof(buf), "SCORE %u", (unsigned)game.score());
+        std::snprintf(alt, sizeof(alt), "%u", (unsigned)game.score());
         drawCenteredFit(g, oy + uiY(135), buf, alt, HUD_TEXT);
-        std::snprintf(buf, sizeof(buf), "HIGH SCORE %u", game.highScore());
-        std::snprintf(alt, sizeof(alt), "HI %u", game.highScore());
+        std::snprintf(buf, sizeof(buf), "HIGH SCORE %u",
+                      (unsigned)game.highScore());
+        std::snprintf(alt, sizeof(alt), "HI %u", (unsigned)game.highScore());
         drawCenteredFit(g, oy + uiY(152), buf, alt,
                         game.score() >= game.highScore() && game.score() > 0
                             ? g2::makeColor(255, 230, 120)
